+++ src/main/java/com/elduin/planets/client/PlanetsClient.java
@@ -11,5 +11,6 @@ public final class PlanetsClient {
 
 	public static void init() {
 		Hooks.openLaunchScreen = () -> Minecraft.getInstance().setScreen(new LaunchScreen());
+		TestWorldButton.register();
 	}
 }